翻转整数算法修改3

+++ app/src/main/java/com/kbein/algorithm/ReverseInteger/IntegerReverse.java
@@ -23,10 +23,10 @@ import android.util.Log;
 public class IntegerReverse {
 
 
-    public static long reverse(Long x) {
-        long result = 0;
+    public static int reverse(int x) {
+        int result = 0;
         while (x !=0){
-            int pop = (int) (x % 10);
+            int pop = x % 10;
             x /=10;
 
             if (result > Integer.MAX_VALUE/10 || (result == Integer.MAX_VALUE/10 && pop > 7)) return 0;
